utils: json_readeer: Fix on 32bit
The int_val has to be long long so that it works correctly on 32bit.

And actually we need to unsigned variant as well so that we can
correctly store 64bit unsigned values. However that will be slightly
more complicated change.

+++ include/utils/gp_json_reader.h
@@ -52,7 +52,7 @@ typedef struct gp_json_val {
 	/** An union to store the parsed value into. */
 	union {
 		int val_bool;
-		long val_int;
+		long long val_int;
 		const char *val_str;
 	};
